... The cable had passed us by; the dish was the only hope, and eventually we were all forced to turn to it.  By the summer of '85, the valley had more satellite dishes per capita than an Eskimo village on the north slope of Alaska.
Mine was one of the last to go in.  I had been nervous from the start about
the hazards of too much input, which is a very real problem with these
things.  Watching TV becomes a full-time job when you can scan 200 channels
all day and all night and still have the option of punching Night Dreams
into the video machine, if the rest of the world seems dull.
-- Hunter Thompson, "Full-time scrambling", _Generation of Swine_