Stayed in bed all morning just to pass the time, There's something wrong here, there can be no more denying, One of us is changing, or maybe we just stopped trying,
And it's too late, baby, now, it's too late,
Though we really did try to make it,
Something inside has died and I can't hide and I just can't fake it...

It used to be so easy living here with you,
You were light and breezy and I knew just what to do
Now you look so unhappy and I feel like a fool.

There'll be good times again for me and you,
But we just can't stay together, don't you feel it too?
But I'm glad for what we had and that I once loved you...

But it's too late baby...
It's too late, now darling, it's too late...
		-- Carol King, "Tapestry"