Fortune Documents the Great Legal Decisions:
We think that we may take judicial notice of the fact that the term "bitch"
may imply some feeling of endearment when applied to a female of the canine
species but that it is seldom, if ever, so used when applied to a female
of the human race. Coming as it did, reasonably close on the heels of two
revolver shots directed at the person of whom it was probably used, we think
it carries every reasonable implication of ill-will toward that person.
		-- Smith v. Moran, 193 N.E. 2d 466.