Inadmissible:  Not competent to be considered.  Said of certain kinds of testimony which juries are supposed to be unfit to be entrusted with, and which judges, therefore, rule out, even of proceedings before themselves alone.  Hearsay evidence is inadmissible because the person quoted was unsworn and is not before the court for examination; yet most momentous actions, military, political, commercial and of every other kind, are daily undertaken on hearsay evidence.  There is no religion in the world that has any other basis than hearsay evidence.  Revelation is hearsay evidence; that the Scriptures are the word of God we have only the testimony of men long dead whose identy is not clearly established and who are not known to have been sworn in any sense.  Under the rules of evidence as they now exist in this country, no single assertion in the Bible has in its support any evidence admissible in a court of law...
But as records of courts of justice are admissible, it can easily be proved
that powerful and malevolent magicians once existed and were a scourge to
mankind.  The evidence (including confession) upon which certain women
were convicted of witchcraft and executed was without a flaw; it is still
unimpeachable.  The judges' decisions based on it were sound in logic and
in law.  Nothing in any existing court was ever more thoroughly proved than
the charges of witchcraft and sorcery for which so many suffered death.
If there were no witches, human testimony and human reason are alike
destitute of value.  --Ambrose Bierce