What I Did During My Fall Semester On the first day of my fall semester, I got up. Then I went to the library to find a thesis topic. Then I hung out in front of the Dover.
On the second day of my fall semester, I got up.
Then I went to the library to find a thesis topic.
Then I hung out in front of the Dover.

On the third day of my fall semester, I got up.
Then I went to the library to find a thesis topic.
I found a thesis topic:
	How to keep people from hanging out in front of the Dover.
		-- Sister Mary Elephant, "Student Statement for Black Friday"