Men's skin is different from women's skin.  It is usually bigger, and it has more snakes tattooed on it.  Also, if you examine a woman's skin very closely, inch by inch, starting at her shapely ankles, then gently tracing the slender curve of her calves, then moving up to her ...
[EDITOR'S NOTE: To make room for news articles about important world events
such as agriculture, we're going to delete the next few square feet of the
woman's skin.  Thank you.]

... until finally the two of you are lying there, spent, smoking your
cigarettes, and suddenly it hits you: Human skin is actually made up of
billions of tiny units of protoplasm, called "cells"!  And what is even more
interesting, the ones on the outside are all dying!  This is a fact.  Your
skin is like an aggressive modern corporation, where the older veteran
cells, who have finally worked their way to the top and obtained offices
with nice views, are constantly being shoved out the window head first,
without so much as a pension plan, by younger hotshot cells moving up from
below.
		-- Dave Barry, "Saving Face"