Got a wife and kids in Baltimore Jack, I went out for a ride and never came back. Like a river that don't know where it's flowing, I took a wrong turn and I just kept going.
	Everybody's got a hungry heart.
	Everybody's got a hungry heart.
	Lay down your money and you play your part,
	Everybody's got a hungry heart.

I met her in a Kingstown bar,
We fell in love, I knew it had to end.
We took what we had and we ripped it apart,
Now here I am down in Kingstown again.

Everybody needs a place to rest,
Everybody wants to have a home.
Don't make no difference what nobody says,
Ain't nobody likes to be alone.
		-- Bruce Springsteen, "Hungry Heart"