****  GROWTH CENTER REPAIR SERVICE
For those who have had too much of Esalen, Topanga, and Kairos. Tired of
being genuine all the time?  Would you like to learn how to be a little
phony again?  Have you disclosed so much that you're beginning to avoid
people? Have you touched so many people that they're all beginning to
feel the same? Like to be a little dependent? Are perfect orgasms
beginning to bore you? Would you like, for once, not to express a
feeling?  Or better yet, not be in touch with it at all?  Come to us.  We
promise to relieve you of the burden of your great potential.