Oh give me your pity! I'm on a committee,			We attend and amend Which means that from morning		And contend and defend 	to night,			Without a conclusion in sight.
We confer and concur,
We defer and demur,			We revise the agenda
And reiterate all of our thoughts.	With frequent addenda
					And consider a load of reports.

We compose and propose,
We suppose and oppose,			But though various notions
And the points of procedure are fun;	Are brought up as motions,
					There's terribly little gets done.

We resolve and absolve;
But we never dissolve,
Since it's out of the question for us
To bring our committee
To end like this ditty,
Which stops with a period, thus.
		-- Leslie Lipson, "The Committee"