Q: They just announced on the radio that Dan Quayle was picked as the Republican V.P. candidate.  Should I post?
A: Of course.  The net can reach people in as few as 3 to 5 days.  It's
the perfect way to inform people about such news events long after the
broadcast networks have covered them.  As you are probably the only person
to have heard the news on the radio, be sure to post as soon as you can.

-- Brad Templeton, _Emily Postnews Answers Your Questions on Netiquette_