Come live with me and be my love, And we will some new pleasures prove Of golden sands and crystal brooks With silken lines, and silver hooks. There's nothing that I wouldn't do If you would be my POSSLQ.
You live with me, and I with you,
And you will be my POSSLQ.
I'll be your friend and so much more;
That's what a POSSLQ is for.

And everything we will confess;
Yes, even to the IRS.
Some day on what we both may earn,
Perhaps we'll file a joint return.
You'll share my pad, my taxes, joint;
You'll share my life - up to a point!
And that you'll be so glad to do,
Because you'll be my POSSLQ.