Q: Somebody just posted that Roman Polanski directed Star Wars.  What should I do?
A: Post the correct answer at once!  We can't have people go on believing
that!  Very good of you to spot this.  You'll probably be the only one to
make the correction, so post as soon as you can.  No time to lose, so
certainly don't wait a day, or check to see if somebody else has made the
correction.

And it's not good enough to send the message by mail.  Since you're the
only one who really knows that it was Francis Coppola, you have to inform
the whole net right away!

-- Brad Templeton, _Emily Postnews Answers Your Questions on Netiquette_