Fortune's Guide to Freshman Notetaking:
WHEN THE PROFESSOR SAYS:			YOU WRITE:

Probably the greatest quality of the poetry	John Milton -- born 1608
of John Milton, who was born in 1608, is the
combination of beauty and power.  Few have
excelled him in the use of the English language,
or for that matter, in lucidity of verse form,
'Paradise Lost' being said to be the greatest
single poem ever written."

Current historians have come to			Most of the problems that now
doubt the complete advantageousness		face the United States are
of some of Roosevelt's policies...		directly traceable to the
						bungling and greed of President
						Roosevelt.

... it is possible that we simply do		Professor Mitchell is a
not understand the Russian viewpoint...		communist.