*** DO YOU HAVE A RESTLESS URGE TO PROGRAM? *** Do you want the instant respect that comes from being able to use technical terms that nobody understands?  Do you want to strike fear and loathing into the hearts of DP managers everywhere?  If so, then let the Famous Programmers' School lead you on... into the world of professional computer programming.
	*** IS PROGRAMMING FOR YOU? ***
Programming is not for everyone.  But, if you have the desire to learn, we can
help you get started.  All you need is the Famous Programmers' Course and
enough money to keep those lessons coming month after month.

	*** TAKE OUR FREE APTITUDE TEST ***
To help determine if you are qualified to be a programmer, take a moment to
try this simple test:
	(1) Write down the numbers from zero to nine and the first six letters
		of the alphabet (Hint: 0123456789ABCDEF).
	(2) Whose picture is on the back of a twenty-dollar bill?
	(3) What is the state capital of Idaho?
If you managed to read all three questions without wondering why we asked
them, you may have a future as a computer programmer.