Ever wonder why fire engines are red?
Because newspapers are read too.
Two and Two is four.
Four and four is eight.
Eight and four is twelve.
There are twelve inches in a ruler.
Queen Mary was a ruler.
Queen Mary was a ship.
Ships sail the sea.
There are fishes in the sea.
Fishes have fins.
The Finns fought the Russians.
Russians are red.
Fire engines are always rush'n.
Therefore fire engines are red.