After two or three weeks of this madness, you begin to feel As One with the man who said, "No news is good news." In twenty-eight papers, only the rarest kind of luck will turn up more than two or three articles of any interest...  but even then the interest items are usually buried deep around paragraph 16 on the jump (or "Cont.  on ...") page...
The Post will have a story about Muskie making a speech in Iowa.  The
Star will say the same thing, and the Journal will say nothing at all.
But the Times might have enough room on the jump page to include a line
or so that says something like: "When he finished his speech, Muskie
burst into tears and seized his campaign manager by the side of the neck.
They grappled briefly, but the struggle was kicked apart by an oriental
woman who seemed to be in control."

Now that's good journalism.  Totally objective; very active and straight
to the point.
		-- Hunter S. Thompson, "Fear and Loathing '72"