THE LESSER-KNOWN PROGRAMMING LANGUAGES #14 -- VALGOL
	VALGOL is enjoying a dramatic surge of popularity across the
industry.  VALGOL commands include REALLY, LIKE, WELL, and Y*KNOW.
Variables are assigned with the =LIKE and =TOTALLY operators.  Other
operators include the "California booleans", AX and NOWAY.  Loops are
accomplished with the FOR SURE construct.  A simple example:

	LIKE, Y*KNOW(I MEAN)START
	IF PIZZA	=LIKE BITCHEN AND
	GUY		=LIKE TUBULAR AND
	VALLEY GIRL	=LIKE GRODY**MAX(FERSURE)**2
	THEN
		FOR I =LIKE 1 TO OH*MAYBE 100
			DO*WAH - (DITTY**2); BARF(I)=TOTALLY GROSS(OUT)
		SURE
	LIKE, BAG THIS PROGRAM; REALLY; LIKE TOTALLY(Y*KNOW); IM*SURE
	GOTO THE MALL

	VALGOL is also characterized by its unfriendly error messages.  For
example, when the user makes a syntax error, the interpreter displays the
message GAG ME WITH A SPOON!  A successful compile may be termed MAXIMALLY
AWESOME!