To understand this important story, you have to understand how the telephone company works.  Your telephone is connected to a local computer, which is in turn connected to a regional computer, which is in turn connected to a loudspeaker the size of a garbage truck on the lawn of Edna A. Bargewater of Lawrence, Kan.
Whenever you talk on the phone, your local computer listens in.  If it
suspects you're going to discuss an intimate topic, it notifies the computer
above it, which listens in and decides whether to alert the one above it,
until finally, if you really humiliate yourself, maybe break down in tears
and tell your closest friend about a sordid incident from your past
involving a seedy motel, a neighbor's spouse, an entire religious order, a
garden hose and six quarts of tapioca pudding, the top computer feeds your
conversation into Edna's loudspeaker, and she and her friends come out on
the porch to listen and drink gin and laugh themselves silly.
		-- Dave Barry, "Won't It Be Just Great Owning Our Own Phones?"