DE:  The Soviets seem to have difficulty implementing modern technology.      Would you comment on that?
Belenko:  Well, let's talk about aircraft engine lifetime.  When I flew the
	  MiG-25, its engines had a total lifetime of 250 hours.

DE:  Is that mean-time-between-failure?

Belenko:  No, the engine is finished; it is scrapped.

DE:  You mean they pull it out and throw it away, not even overhauling it?

Belenko:  That is correct.  Overhaul is too expensive.

DE:  That is absurdly low by free world standards.

Belenko:  I know.
-- an interview with Victor Belenko, MiG-25 fighter pilot who defected in 1976
   "Defense Electronics", Vol 20, No. 6, pg. 102