THE "FUN WITH USENET" MANIFESTO Very little happens on Usenet without some sort of response from some other reader.  Fun With Usenet postings are no exception.  Since there are some who might question the rationale of some of the excerpts included therein, I have written up a list of guidelines that sum up the philosophy behind these postings.
	One.  I never cut out words in the middle of a quote without a VERY
good reason, and I never cut them out without including ellipses.  For
instance, "I am not a goob" might become "I am ... a goob", but that's too
mundane to bother with.  "I'm flame proof" might (and has) become
"I'm ...a... p...oof" but that's REALLY stretching it.

	Two.  If I cut words off the beginning or end of a quote, I don't
put ellipses, but neither do I capitalize something that wasn't capitalized
before the cut. "I don't think that the Church of Ubizmo is a wonderful
place" would turn into "the Church of Ubizmo is a wonderful place".  Imagine
the posting as a tape-recording of the poster's thoughts.  If I can set
up the quote via fast-forwarding and stopping the tape, and without splicing,
I don't put ellipses in.  And by the way, I love using this mechanism for
turning things around.  If you think something stinks, say so - don't say you
don't think it's wonderful.   ...
-- D. J. McCarthy (dmccart@cadape.UUCP)