THE LESSER-KNOWN PROGRAMMING LANGUAGES #8: LAIDBACK
This language was developed at the Marin County Center for T'ai Chi,
Mellowness and Computer Programming (now defunct), as an alternative to
the more intense atmosphere in nearby Silicon Valley.

The center was ideal for programmers who liked to soak in hot tubs while
they worked.  Unfortunately few programmers could survive there because the
center outlawed Pizza and Coca-Cola in favor of Tofu and Perrier.

Many mourn the demise of LAIDBACK because of its reputation as a gentle and
non-threatening language since all error messages are in lower case.  For
example, LAIDBACK responded to syntax errors with the message:

	"i hate to bother you, but i just can't relate to that.  can
	you find the time to try it again?"