One bright Sunday morning, in the shadows of the steeple, By the Relief Office, I seen my people; As they stood there hungry, I stood there whistling, This land was made for you and me.
Nobody living can ever stop me,
As I go walking that freedom highway;
Nobody living can ever make me turn back,
This land was made for you and me.

As I went walking, I saw a sign there,
And on the sign it said: "No Trespassing."
But on the other side, it didn't say nothing,
That side was made for you and me.
		-- Woody Guthrie, "This Land Is Your Land" (verses 4, 6, 7)
	[If you ever wondered why Arlo was so anti-establishment when his dad
	 wrote such wonderful patriotic songs, the answer is that you haven't
	 heard all of Woody's songs]