Well, we're big rock singers, we've got golden fingers, And we're loved everywhere we go. We sing about beauty, and we sing about truth, At ten thousand dollars a show. We take all kind of pills to give us all kind of thrills, But the thrill we've never known, Is the thrill that'll get'cha, when you get your picture, On the cover of the Rolling Stone.
I got a freaky old lady, name of Cole King Katie,
Who embroiders on my jeans.
I got my poor old gray-haired daddy,
Drivin' my limousine.
Now it's all designed, to blow our minds,
But our minds won't be really be blown;
Like the blow that'll get'cha, when you get your picture,
On the cover of the Rolling Stone.

We got a lot of little, teen-aged, blue-eyed groupies,
Who'll do anything we say.
We got a genuine Indian guru, that's teachin' us a better way.
We got all the friends that money can buy,
So we never have to be alone.
And we keep gettin' richer, but we can't get our picture,
On the cover of the Rolling Stone.
		-- Dr. Hook and the Medicine Show
		[As a note, they eventually DID make the cover of RS. Ed.]